Added approved change to clarify intent is not to chill speech
"CppCon strives to be a respectful, kind, diverse, and inclusive environment. Participants are expected to not be offensive or disrespectful toward others. They are also encouraged to not be too quick to take offense when none is intended but assume good faith when interacting with persons of different cultures, life experiences, and backgrounds who may have diverse ways of expression. The CoC exists to ensure we have a welcoming atmosphere, not to chill communication by trying to control how people express themselves as long as they are respectful."